Refatoração T04 seguindo MVVM
