+++ src/screens/CreateParty/CreatePartyViewModel.ts
@@ -1,0 +1,108 @@
+import { useState } from "react";
+import { Party } from "../../@types/Party";
+
+// O ViewModel é apenas um Custom Hook que recebe a navegação para conseguir mudar de tela
+export const useCreatePartyViewModel = (navigation: any) => {
+  const [isModalVisible, setModalVisible] = useState(false);
+  const [nomeParty, setNomeParty] = useState("");
+  const [dataRevelacao, setDataRevelacao] = useState<Date | undefined>(undefined);
+  const [valorMinimo, setValorMinimo] = useState("");
+  const [valorMaximo, setValorMaximo] = useState("");
+  const [errors, setErrors] = useState({ nome: "", data: "", valores: "" });
+
+  // Funções de atualização que já limpam os erros automaticamente
+  const updateNomeParty = (text: string) => {
+    setNomeParty(text);
+    if (errors.nome) setErrors((prev) => ({ ...prev, nome: "" }));
+  };
+
+  const updateDataRevelacao = (date: Date) => {
+    setDataRevelacao(date);
+    if (errors.data) setErrors((prev) => ({ ...prev, data: "" }));
+  };
+
+  const updateValorMinimo = (text: string) => {
+    setValorMinimo(text);
+    if (errors.valores) setErrors((prev) => ({ ...prev, valores: "" }));
+  };
+
+  const updateValorMaximo = (text: string) => {
+    setValorMaximo(text);
+    if (errors.valores) setErrors((prev) => ({ ...prev, valores: "" }));
+  };
+
+  // Funções de Navegação e Modal
+  const handleBackPress = () => setModalVisible(true);
+  const cancelExit = () => setModalVisible(false);
+  const confirmExit = () => {
+    setModalVisible(false);
+    navigation.goBack();
+  };
+
+  const parseCurrency = (value: string) => {
+    if (!value) return 0;
+    return parseFloat(value.replace(/\./g, "").replace(",", "."));
+  };
+
+  // Validação e Regra de Negócio (T04)
+  const handleCriarParty = () => {
+    let isValid = true;
+    let newErrors = { nome: "", data: "", valores: "" };
+
+    if (!nomeParty.trim()) {
+      newErrors.nome = "O nome da Party é obrigatório.";
+      isValid = false;
+    }
+
+    if (!dataRevelacao) {
+      newErrors.data = "Selecione a data da revelação.";
+      isValid = false;
+    }
+
+    const numMin = parseCurrency(valorMinimo);
+    const numMax = parseCurrency(valorMaximo);
+
+    if (valorMinimo === "" || valorMaximo === "") {
+      newErrors.valores = "Preencha o valor mínimo e máximo.";
+      isValid = false;
+    } else if (numMax < numMin) {
+      newErrors.valores = "O valor máximo não pode ser menor que o mínimo.";
+      isValid = false;
+    }
+
+    setErrors(newErrors);
+
+    if (isValid) {
+      const novaParty: Party = {
+        id: Math.random().toString(36).substring(2, 10),
+        name: nomeParty,
+        eventDate: dataRevelacao!.toISOString(),
+        minPrice: numMin,
+        maxPrice: numMax,
+        maxParticipants: 10,
+        status: "Aguardando Sorteio",
+      };
+
+      console.log("SUCESSO (ViewModel)! Party Criada:", novaParty);
+      navigation.navigate("PartyCreated", { party: novaParty });
+    }
+  };
+
+  // Retornamos apenas o que a View precisa enxergar e utilizar
+  return {
+    nomeParty,
+    updateNomeParty,
+    dataRevelacao,
+    updateDataRevelacao,
+    valorMinimo,
+    updateValorMinimo,
+    valorMaximo,
+    updateValorMaximo,
+    errors,
+    isModalVisible,
+    cancelExit,
+    handleBackPress,
+    confirmExit,
+    handleCriarParty,
+  };
+};
